requirements.txt
faiss-cpu
openai
numpy
python-dotenv	
pymupdf
tenacity
ragas
langsmith
langchain
langchain-openai
scikit-learn
sentence-transformers
ruff
black
rank-bm25